feat: Refactor ControlPanel into two components
Splits the ControlPanel functionality into two distinct components:
- `ControlPanel.ts`: Retains only the core "Controls" tab (population, gold, troop/worker ratios, attack ratio, investment rate).
- `ControlPanel2.ts`: Houses the "Bombers", "Build", "Nukes", "Units", and "Options" tabs.

This refactoring improves modularity and separation of concerns.

Fixes: Corrected an issue where the build menu was not appearing correctly after right-clicking a tile due to the `ShowBuildMenuInControlPanelEvent` not being handled in the new `ControlPanel2.ts`. The event listener has been re-added to `ControlPanel2.ts` to ensure proper functionality.

refractor notifications and controls panels
