+++ src/client/graphics/layers/ChatDisplay.ts
@@ -127,7 +127,7 @@ export class ChatDisplay extends LitElement implements Layer {
       <div
         class="${this._hidden
           ? "w-fit px-[10px] py-[5px]"
-          : ""} rounded-md bg-black bg-opacity-60 relative max-h-[30vh] flex flex-col-reverse overflow-y-auto w-full lg:bottom-2.5 lg:right-2.5 z-50 lg:max-w-[30vw] lg:w-full lg:w-auto"
+          : ""} rounded-md bg-black bg-opacity-60 relative flex flex-col-reverse overflow-y-auto w-full h-full"
         style="pointer-events: auto"
       >
         <div>
