+++ src/client/index.html
@@ -217,18 +217,17 @@
     </div>
 
     <div
-      class="bottom-0 w-full flex-col-reverse sm:flex-row z-50 md:w-[320px]"
+      class="bottom-0 w-full flex z-50"
       style="position: fixed; pointer-events: none"
     >
-      <div
-        class="w-full sm:w-2/3 sm:fixed sm:right-0 sm:bottom-0 sm:flex justify-end"
-        style="pointer-events: none"
-      >
+      <control-panel class="w-1/4" style="pointer-events: auto"></control-panel>
+      <control-panel2
+        class="w-1/2"
+        style="pointer-events: auto"
+      ></control-panel2>
+      <div class="w-1/4 flex flex-col" style="pointer-events: auto">
         <chat-display></chat-display>
         <events-display></events-display>
-      </div>
-      <div class="w-full sm:w-1/3 md:max-w-72" style="pointer-events: auto">
-        <control-panel></control-panel>
       </div>
     </div>
